Fix truncation of rom doubles written to program bin.
When the ROM double array is written to the xc_program.bin file, the default ostream conversion of double is used which results in truncation of double values.    For example, 123456789.123456789 will be stored as 123456700 in the xc_program.bin.   This update uses the existing ToString(double) routine to convert the double to string for consistent conversion.

Updated the unit test for various double values.   To see the before fix, run updated unit test and note the results of test 28 before applying the XenonCode.hpp change.   Checked under g++/linux and msvc/windows.