Incorrect size passed to array.reserve for array append.
While tracing code, noticed a type-o related to the array.reserve call for an array append.   Uses the args.size() when it should be the new size of the array (current + args).   No functionality impact, just small memory perf.    Appears some common code could be used, but outside the scope of this change.

Tested under g++/linux and msvc/window.