Remove undefines for Windows C++
Removed the region UNDEFS and no issues compiling with msvc/windows (and g++/linux).    I'm on the current Visual C++ 2022 (under Visual Studio).    Are you still able to repo an issue with the UNDEFS region removed?   If yes, how are you building?    (Not sure if this is an old issue or still current.)