Switch to implicit vector.at checking vs explicit code.
Changed a single method to leverage the vector.at method that validates the range and throws std::out_of_range on failure.   Small update to verify switching to using vector.at is okay for the rest of the code.   It will require the host to catch the std::out_of_range exception and handle in the current RuntimeError exception catch.   If accepted, will change the rest of the code.

Also fixed up the RAM_VAR_TEXT handler to use the new ToDouble routine.   (Also need to scan the rest of the code for other instances, but trying to keep this change small.)